containers/reflection/requirements.txt
fastapi
uvicorn
httpx
mcp[sse]>=1.2.0
python-dotenv
google-cloud-storage
openai
pandas
